requirements.txt
Django>=5.0.6,<5.1
djangorestframework>=3.15.2,<3.16
django-filter
drf-spectacular>=0.26.5,<0.27
psycopg2>=2.9.9,<3
numpy
scipy
djangorestframework-simplejwt